extract sentry stuff to own module
